requirements.txt
fastapi
uvicorn
streamlit
tensorflow
opencv-python
numpy
pandas
scikit-learn
pillow
requests
